refactor: 엔티티 읽어들이는 메소드 클래스 분리
